PDASHOSS01-123 Defer pydantic-ai out of runtime/mcp.py's import path
runtime/mcp.py imported pydantic-ai at module scope
(`from pydantic_ai import exceptions` and
`from pydantic_ai.toolsets import WrapperToolset`), so every importer of
this module — reached at Django URLconf load time — pulled pydantic-ai and
its transitive tree into the process. The rest of the assistant runtime
(runtime/llm.py, tasks.py, build_toolset) already defers pydantic-ai to
function scope; this module was the exception.

WrapperToolset was at module scope only because ResilientToolset subclasses
it, and a class statement cannot defer its base. Move ResilientToolset (and
the _CONTROL_FLOW_EXCEPTIONS table that also needs pydantic-ai) into a new
sibling module runtime/resilient_toolset.py, import it lazily inside
build_toolset (alongside the existing local MCPToolset import), and add a
module-level __getattr__ so `mcp.ResilientToolset` still resolves lazily for
callers and tests without re-triggering the startup import.
